4.2 Move `extract_utxo_outputs_from_wallet_state` which uses the `StoredOutput` in `src/storage/storage_trait.rs` put it in the most logical place
