+++ src/bin/scanner.rs
@@ -103,18 +103,13 @@
 
 #[cfg(feature = "grpc")]
 use clap::Parser;
-#[cfg(all(feature = "grpc", feature = "storage"))]
-use lightweight_wallet_libs::storage::{OutputStatus, StoredOutput};
+
 #[cfg(feature = "grpc")]
 use lightweight_wallet_libs::{
     common::format_number,
     data_structures::{
-        block::Block,
-        payment_id::PaymentId,
-        transaction::TransactionDirection,
-        transaction_output::LightweightTransactionOutput,
-        types::{CompressedCommitment, PrivateKey},
-        wallet_transaction::WalletState,
+        block::Block, payment_id::PaymentId, transaction::TransactionDirection,
+        types::CompressedCommitment, wallet_transaction::WalletState,
     },
     errors::LightweightWalletResult,
     scanning::{
@@ -410,410 +405,6 @@ pub enum ScanResult {
     Interrupted(WalletState),
 }
 
-/// Extract UTXO data from blockchain outputs and create StoredOutput objects
-#[cfg(all(feature = "grpc", feature = "storage"))]
-fn extract_utxo_outputs_from_wallet_state(
-    wallet_state: &WalletState,
-    scan_context: &ScanContext,
-    wallet_id: u32,
-    block_outputs: &[LightweightTransactionOutput],
-    block_height: u64,
-) -> LightweightWalletResult<Vec<StoredOutput>> {
-    use lightweight_wallet_libs::data_structures::transaction::TransactionDirection;
-
-    let mut utxo_outputs = Vec::new();
-
-    // Get inbound transactions from this specific block
-    let block_transactions: Vec<_> = wallet_state
-        .transactions
-        .iter()
-        .filter(|tx| {
-            tx.block_height == block_height
-                && tx.transaction_direction == TransactionDirection::Inbound
-        })
-        .collect();
-
-    for transaction in block_transactions {
-        // Find the corresponding blockchain output
-        if let Some(output_index) = transaction.output_index {
-            if let Some(blockchain_output) = block_outputs.get(output_index) {
-                // Derive spending keys for this output
-                let (spending_key, script_private_key) =
-                    derive_utxo_spending_keys(&scan_context.entropy, output_index as u64)?;
-
-                // Extract script input data and lock height
-                let (input_data, script_lock_height) =
-                    extract_script_data(&blockchain_output.script.bytes)?;
-
-                // Create StoredOutput from blockchain data
-                let stored_output = StoredOutput {
-                    id: None, // Will be set by database
-                    wallet_id,
-
-                    // Core UTXO identification
-                    commitment: blockchain_output.commitment.as_bytes().to_vec(),
-                    hash: compute_output_hash(blockchain_output)?,
-                    value: transaction.value,
-
-                    // Spending keys (derived from entropy)
-                    spending_key: hex::encode(spending_key.as_bytes()),
-                    script_private_key: hex::encode(script_private_key.as_bytes()),
-
-                    // Script and covenant data
-                    script: blockchain_output.script.bytes.clone(),
-                    input_data,
-                    covenant: blockchain_output.covenant.bytes.clone(),
-
-                    // Output features and type
-                    output_type: blockchain_output.features.output_type.clone() as u32,
-                    features_json: serde_json::to_string(&blockchain_output.features).map_err(
-                        |e| {
-                            LightweightWalletError::StorageError(format!(
-                                "Failed to serialize features: {e}"
-                            ))
-                        },
-                    )?,
-
-                    // Maturity and lock constraints
-                    maturity: blockchain_output.features.maturity,
-                    script_lock_height,
-
-                    // Metadata signature components
-                    sender_offset_public_key: blockchain_output
-                        .sender_offset_public_key
-                        .as_bytes()
-                        .to_vec(),
-                    // Note: LightweightSignature only has bytes field, so we use placeholders
-                    // In a full implementation, these would be extracted from the signature structure
-                    metadata_signature_ephemeral_commitment: vec![0u8; 32], // Placeholder
-                    metadata_signature_ephemeral_pubkey: vec![0u8; 32],     // Placeholder
-                    metadata_signature_u_a: if blockchain_output.metadata_signature.bytes.len()
-                        >= 32
-                    {
-                        blockchain_output.metadata_signature.bytes[0..32].to_vec()
-                    } else {
-                        vec![0u8; 32]
-                    },
-                    metadata_signature_u_x: if blockchain_output.metadata_signature.bytes.len()
-                        >= 64
-                    {
-                        blockchain_output.metadata_signature.bytes[32..64].to_vec()
-                    } else {
-                        vec![0u8; 32]
-                    },
-                    metadata_signature_u_y: vec![0u8; 32], // Placeholder
-
-                    // Payment information
-                    encrypted_data: blockchain_output.encrypted_data.as_bytes().to_vec(),
-                    minimum_value_promise: blockchain_output.minimum_value_promise.as_u64(),
-
-                    // Range proof
-                    rangeproof: blockchain_output.proof.as_ref().map(|p| p.bytes.clone()),
-
-                    // Status and spending tracking
-                    status: if transaction.is_spent {
-                        OutputStatus::Spent as u32
-                    } else {
-                        OutputStatus::Unspent as u32
-                    },
-                    mined_height: Some(transaction.block_height),
-                    spent_in_tx_id: if transaction.is_spent {
-                        // Calculate transaction ID from spent block and input index
-                        transaction.spent_in_block.and_then(|spent_block| {
-                            transaction.spent_in_input.map(|spent_input| {
-                                generate_transaction_id(spent_block, spent_input)
-                            })
-                        })
-                    } else {
-                        None
-                    },
-
-                    // Timestamps (will be set by database)
-                    created_at: None,
-                    updated_at: None,
-                };
-
-                utxo_outputs.push(stored_output);
-            }
-        }
-    }
-
-    Ok(utxo_outputs)
-}
-
-/// Extract script input data and script lock height from script bytes
-#[cfg(all(feature = "grpc", feature = "storage"))]
-fn extract_script_data(script_bytes: &[u8]) -> LightweightWalletResult<(Vec<u8>, u64)> {
-    // If script is empty, return empty data
-    if script_bytes.is_empty() {
-        return Ok((Vec::new(), 0));
-    }
-
-    let mut input_data = Vec::new();
-    let mut script_lock_height = 0u64;
-    let mut potential_heights = Vec::new();
-
-    // Parse script bytecode to extract data
-    // This is a simplified parser - in a full implementation, you'd use a proper script interpreter
-    let mut i = 0;
-    while i < script_bytes.len() {
-        match script_bytes[i] {
-            // OP_PUSHDATA opcodes (0x01-0x4b) - extract the data being pushed
-            0x01..=0x4b => {
-                let data_len = script_bytes[i] as usize;
-                i += 1;
-                if i + data_len <= script_bytes.len() {
-                    let data = script_bytes[i..i + data_len].to_vec();
-
-                    // Check if this data might be input data (execution stack data)
-                    // Input data is typically non-zero and has meaningful structure
-                    if !data.iter().all(|&b| b == 0) && !data.is_empty() {
-                        // Prefer larger, more structured data as input data
-                        if input_data.is_empty() || data.len() > input_data.len() {
-                            input_data = data.clone();
-                        }
-                    }
-
-                    // Check if this could be a height value (4 or 8 bytes)
-                    if data.len() == 4 || data.len() == 8 {
-                        let height = if data.len() == 4 {
-                            u32::from_le_bytes(data.clone().try_into().unwrap_or([0; 4])) as u64
-                        } else {
-                            u64::from_le_bytes(data.clone().try_into().unwrap_or([0; 8]))
-                        };
-
-                        // Store as potential height if it looks reasonable
-                        if height > 0 && height < 10_000_000 && height > 100 {
-                            potential_heights.push(height);
-                        }
-                    }
-
-                    i += data_len;
-                } else {
-                    break; // Malformed script
-                }
-            }
-
-            // OP_PUSHDATA1 (0x4c) - next byte is data length
-            0x4c => {
-                if i + 1 < script_bytes.len() {
-                    let data_len = script_bytes[i + 1] as usize;
-                    i += 2;
-                    if i + data_len <= script_bytes.len() {
-                        let data = script_bytes[i..i + data_len].to_vec();
-                        if !data.iter().all(|&b| b == 0)
-                            && !data.is_empty()
-                            && (input_data.is_empty() || data.len() > input_data.len())
-                        {
-                            input_data = data.clone();
-                        }
-
-                        // Check for height values
-                        if data.len() == 4 || data.len() == 8 {
-                            let height = if data.len() == 4 {
-                                u32::from_le_bytes(data.clone().try_into().unwrap_or([0; 4])) as u64
-                            } else {
-                                u64::from_le_bytes(data.clone().try_into().unwrap_or([0; 8]))
-                            };
-
-                            if height > 0 && height < 10_000_000 && height > 100 {
-                                potential_heights.push(height);
-                            }
-                        }
-
-                        i += data_len;
-                    } else {
-                        break;
-                    }
-                } else {
-                    break;
-                }
-            }
-
-            // OP_PUSHDATA2 (0x4d) - next 2 bytes are data length (little-endian)
-            0x4d => {
-                if i + 2 < script_bytes.len() {
-                    let data_len =
-                        u16::from_le_bytes([script_bytes[i + 1], script_bytes[i + 2]]) as usize;
-                    i += 3;
-                    if i + data_len <= script_bytes.len() {
-                        let data = script_bytes[i..i + data_len].to_vec();
-                        if !data.iter().all(|&b| b == 0)
-                            && !data.is_empty()
-                            && data.len() <= 256
-                            && (input_data.is_empty() || data.len() > input_data.len())
-                        {
-                            input_data = data;
-                        }
-                        i += data_len;
-                    } else {
-                        break;
-                    }
-                } else {
-                    break;
-                }
-            }
-
-            // OP_CHECKHEIGHTVERIFY or similar time-lock opcodes
-            // In Tari, this might be represented differently, but we'll look for patterns
-            0x65..=0x6a => {
-                // Time-lock related opcodes - try to extract height from following bytes
-                if i + 8 < script_bytes.len() {
-                    // Try to read the next 8 bytes as a u64 height
-                    let height_bytes = &script_bytes[i + 1..i + 9];
-                    if let Ok(height_array) = height_bytes.try_into() {
-                        let height = u64::from_le_bytes(height_array);
-                        if height > 0 && height < 10_000_000 && height > 100 {
-                            script_lock_height = height;
-                        }
-                    }
-                }
-                i += 1;
-            }
-
-            // OP_NOP and other common opcodes that might precede time locks
-            0x61 => {
-                // OP_NOP - check if followed by height data
-                if i + 9 < script_bytes.len() {
-                    let height_bytes = &script_bytes[i + 1..i + 9];
-                    if let Ok(height_array) = height_bytes.try_into() {
-                        let height = u64::from_le_bytes(height_array);
-                        if height > 0 && height < 10_000_000 && height > 100 {
-                            potential_heights.push(height);
-                        }
-                    }
-                }
-                i += 1;
-            }
-
-            // All other opcodes
-            _ => {
-                i += 1;
-            }
-        }
-    }
-
-    // If no explicit script lock height was found, use the best candidate from potential heights
-    if script_lock_height == 0 && !potential_heights.is_empty() {
-        // Sort potential heights and pick the most reasonable one
-        potential_heights.sort();
-
-        // Prefer heights that are in the typical blockchain range
-        for &height in &potential_heights {
-            if height > 1000 && height < 1_000_000 {
-                script_lock_height = height;
-                break;
-            }
-        }
-
-        // If no reasonable height found, use the smallest positive one
-        if script_lock_height == 0 {
-            script_lock_height = potential_heights[0];
-        }
-    }
-
-    // Additional heuristic: scan for 8-byte sequences that look like heights
-    if script_lock_height == 0 && script_bytes.len() >= 8 {
-        for chunk_start in 0..=(script_bytes.len() - 8) {
-            if let Ok(height_bytes) = script_bytes[chunk_start..chunk_start + 8].try_into() {
-                let potential_height = u64::from_le_bytes(height_bytes);
-                // More restrictive check for reasonable block heights
-                if potential_height > 1000 && potential_height < 1_000_000 {
-                    script_lock_height = potential_height;
-                    break;
-                }
-            }
-        }
-    }
-
-    Ok((input_data, script_lock_height))
-}
-
-/// Generate a deterministic transaction ID from block height and input index
-#[cfg(all(feature = "grpc", feature = "storage"))]
-fn generate_transaction_id(block_height: u64, input_index: usize) -> u64 {
-    // Create a deterministic transaction ID by combining block height and input index
-    // This is a simplified approach - in a real implementation, you'd use the actual transaction hash
-    //
-    // Format: [32-bit block_height][32-bit input_index]
-    // This ensures unique IDs while being deterministic and easily debuggable
-
-    // Use the block height as the upper 32 bits and input index as lower 32 bits
-    let tx_id = ((block_height & 0xFFFFFFFF) << 32) | (input_index as u64 & 0xFFFFFFFF);
-
-    // Ensure we don't return 0 (which is often treated as "no transaction")
-    if tx_id == 0 {
-        1
-    } else {
-        tx_id
-    }
-}
-
-/// Derive spending keys for a UTXO output using wallet entropy
-/// For view-key mode (entropy all zeros), returns placeholder keys
-#[cfg(all(feature = "grpc", feature = "storage"))]
-fn derive_utxo_spending_keys(
-    entropy: &[u8; 16],
-    output_index: u64,
-) -> LightweightWalletResult<(PrivateKey, PrivateKey)> {
-    use lightweight_wallet_libs::errors::KeyManagementError;
-    use lightweight_wallet_libs::key_management::key_derivation;
-
-    // Check if we have real entropy or if this is view-key mode
-    let has_real_entropy = entropy != &[0u8; 16];
-
-    if has_real_entropy {
-        // Derive real spending keys using wallet entropy
-        let spending_key_raw = key_derivation::derive_private_key_from_entropy(
-            entropy,
-            "wallet_spending", // Branch for spending keys
-            output_index,
-        )?;
-
-        let script_private_key_raw = key_derivation::derive_private_key_from_entropy(
-            entropy,
-            "script_keys", // Branch for script keys
-            output_index,
-        )?;
-
-        // Convert to PrivateKey type
-        let spending_key =
-            PrivateKey::new(spending_key_raw.as_bytes().try_into().map_err(|_| {
-                KeyManagementError::key_derivation_failed("Failed to convert spending key")
-            })?);
-
-        let script_private_key =
-            PrivateKey::new(script_private_key_raw.as_bytes().try_into().map_err(|_| {
-                KeyManagementError::key_derivation_failed("Failed to convert script private key")
-            })?);
-
-        Ok((spending_key, script_private_key))
-    } else {
-        // View-key mode: use placeholder keys (cannot spend, but can store UTXO structure)
-        let placeholder_key_bytes = [0u8; 32];
-        let spending_key = PrivateKey::new(placeholder_key_bytes);
-        let script_private_key = PrivateKey::new(placeholder_key_bytes);
-
-        Ok((spending_key, script_private_key))
-    }
-}
-
-/// Compute output hash for UTXO identification
-#[cfg(all(feature = "grpc", feature = "storage"))]
-fn compute_output_hash(output: &LightweightTransactionOutput) -> LightweightWalletResult<Vec<u8>> {
-    use blake2::{Blake2b, Digest};
-    use digest::consts::U32;
-
-    // Compute hash of output fields for identification
-    let mut hasher = Blake2b::<U32>::new();
-    hasher.update(output.commitment.as_bytes());
-    hasher.update(output.script.bytes.as_slice());
-    hasher.update(output.sender_offset_public_key.as_bytes());
-    hasher.update(output.minimum_value_promise.as_u64().to_le_bytes());
-
-    Ok(hasher.finalize().to_vec())
-}
-
 /// Core scanning logic - simplified and focused with batch processing
 #[cfg(feature = "grpc")]
 async fn scan_wallet_across_blocks_with_cancellation(
@@ -1102,7 +693,7 @@ async fn scan_wallet_across_blocks_with_cancellation(
                         }
 
                         // Extract and save UTXO data for wallet outputs (works for both seed phrase and view-key modes)
-                        match extract_utxo_outputs_from_wallet_state(
+                        match lightweight_wallet_libs::scanning::extract_utxo_outputs_from_wallet_state(
                             &wallet_state,
                             scan_context,
                             storage_backend.wallet_id.unwrap(),
